implement MainActivity and currentSong activity song control functions
implement next, play/pause and prev buttons.
implement sont title update function on the main and currentSong activities.
Left some bugs.
bug #1: when next/prev are clicked on the notification, the song title on either the currentSong
or MainActivity does not update automatically. It depends on the onResume() activity callback.
bug #2: when play/pause is clicked on the notification, the currentSong and the MainActivity play/pause
button are not updated automatically.
Proposed solution: research on, and resolve using service to activity and vice versa communication mechanisms.